e2e: split and move api e2e tests into it's own package (#4751)
* e2e: move api e2e tests into api package

* move api e2e tests into test-e2e-api target and CI job

* fix lint

(cherry picked from commit f8339e6053fdae1d48c492453de1f31cf403f090)
