+++ .github/workflows/ci.yml
@@ -86,6 +86,7 @@ jobs:
         test-target:
           [
             test-e2e,
+            test-e2e-api,
             test-integration-poller,
             test-e2e-deployments,
             test-e2e-serverless,
