+++ .gitignore
@@ -21,6 +21,7 @@
 /tempodb/encoding/benchmark_block
 private-key.key
 integration/e2e/e2e_integration_test[0-9]*
+integration/e2e/api/e2e_integration_test[0-9]*
 integration/e2e/deployments/e2e_integration_test[0-9]*
 .tempo.yaml
 /tmp
