+++ integration/e2e/api/config-plugin-test.yaml
