+++ integration/e2e/api/cross_cluster_reads_test.go
@@ -1,4 +1,4 @@
-package e2e
+package api
 
 import (
 	"testing"
