+++ integration/e2e/api/query_plugin_test.go
@@ -1,4 +1,4 @@
-package e2e
+package api
 
 import (
 	"context"
